# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Fine-Print Guardian API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthRegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  name?: string;
}

export interface AuthLoginInput {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number | string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  name?: string | null;
  isGuest: boolean;
  /** @nullable */
  createdAt?: string | null;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export type DocumentCategory = typeof DocumentCategory[keyof typeof DocumentCategory];


export const DocumentCategory = {
  subscription: 'subscription',
  lease: 'lease',
  tos: 'tos',
} as const;

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  pending: 'pending',
  analyzing: 'analyzing',
  complete: 'complete',
  error: 'error',
} as const;

/**
 * @nullable
 */
export type DocumentOverallScore = typeof DocumentOverallScore[keyof typeof DocumentOverallScore] | null;


export const DocumentOverallScore = {
  safe: 'safe',
  caution: 'caution',
  red_flag: 'red_flag',
} as const;

export interface Document {
  id: number;
  title: string;
  category: DocumentCategory;
  status: DocumentStatus;
  /** @nullable */
  overallScore?: DocumentOverallScore;
  clauseCount: number;
  redFlagCount: number;
  cautionCount: number;
  safeCount: number;
  uploadedAt: string;
  rawText: string;
  /** @nullable */
  errorMessage?: string | null;
}

export type DocumentInputCategory = typeof DocumentInputCategory[keyof typeof DocumentInputCategory];


export const DocumentInputCategory = {
  subscription: 'subscription',
  lease: 'lease',
  tos: 'tos',
} as const;

export interface DocumentInput {
  /** @minLength 1 */
  title: string;
  category: DocumentInputCategory;
  /** @minLength 1 */
  rawText: string;
}

export type DocumentWithClausesCategory = typeof DocumentWithClausesCategory[keyof typeof DocumentWithClausesCategory];


export const DocumentWithClausesCategory = {
  subscription: 'subscription',
  lease: 'lease',
  tos: 'tos',
} as const;

export type DocumentWithClausesStatus = typeof DocumentWithClausesStatus[keyof typeof DocumentWithClausesStatus];


export const DocumentWithClausesStatus = {
  pending: 'pending',
  analyzing: 'analyzing',
  complete: 'complete',
  error: 'error',
} as const;

/**
 * @nullable
 */
export type DocumentWithClausesOverallScore = typeof DocumentWithClausesOverallScore[keyof typeof DocumentWithClausesOverallScore] | null;


export const DocumentWithClausesOverallScore = {
  safe: 'safe',
  caution: 'caution',
  red_flag: 'red_flag',
} as const;

export type ClauseRiskLevel = typeof ClauseRiskLevel[keyof typeof ClauseRiskLevel];


export const ClauseRiskLevel = {
  safe: 'safe',
  caution: 'caution',
  red_flag: 'red_flag',
} as const;

export interface Clause {
  id: number;
  documentId: number;
  text: string;
  categoryTag: string;
  riskLevel: ClauseRiskLevel;
  explanation: string;
  /** @nullable */
  negotiationPrompt?: string | null;
  positionStart: number;
  positionEnd: number;
}

export interface DocumentWithClauses {
  id: number;
  title: string;
  category: DocumentWithClausesCategory;
  status: DocumentWithClausesStatus;
  /** @nullable */
  overallScore?: DocumentWithClausesOverallScore;
  clauseCount: number;
  redFlagCount: number;
  cautionCount: number;
  safeCount: number;
  uploadedAt: string;
  rawText: string;
  /** @nullable */
  errorMessage?: string | null;
  clauses: Clause[];
}

export type ScanStatusStatus = typeof ScanStatusStatus[keyof typeof ScanStatusStatus];


export const ScanStatusStatus = {
  pending: 'pending',
  analyzing: 'analyzing',
  complete: 'complete',
  error: 'error',
} as const;

export interface ScanStatus {
  documentId: number;
  status: ScanStatusStatus;
  progress: number;
  message: string;
}

export interface Watch {
  id: number;
  documentId: number;
  documentTitle: string;
  active: boolean;
  /** @nullable */
  forwardingAddress?: string | null;
  createdAt: string;
  /** @nullable */
  lastCheckedAt?: string | null;
  /** @nullable */
  changesDetected?: boolean | null;
  /** @nullable */
  changeSummary?: string | null;
}

export interface WatchInput {
  documentId: number;
}

export type DashboardDataRiskDistribution = {
  safe: number;
  caution: number;
  red_flag: number;
};

export type DashboardDataCategoryBreakdown = {
  subscription: number;
  lease: number;
  tos: number;
};

export interface DashboardData {
  totalDocuments: number;
  analyzedDocuments: number;
  activeWatches: number;
  riskDistribution: DashboardDataRiskDistribution;
  recentDocuments: Document[];
  categoryBreakdown: DashboardDataCategoryBreakdown;
}

export interface ErrorResponse {
  error: string;
}

export type AuthLogout200 = {
  ok: boolean;
};