# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fine-Print Guardian API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload a document for analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Document too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document with its clauses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document with clauses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentWithClauses"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/scan:
    post:
      operationId: scanDocument
      tags: [documents]
      summary: Trigger clause analysis for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: Analysis started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already analyzing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/scan-status:
    get:
      operationId: getDocumentScanStatus
      tags: [documents]
      summary: Poll analysis progress for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Current scan status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documents/{id}/clauses:
    get:
      operationId: listDocumentClauses
      tags: [clauses]
      summary: List all analyzed clauses for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of clauses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clause"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    AuthRegisterInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 8
        name:
          type: string
      required:
        - email
        - password
    AuthLoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          oneOf:
            - type: integer
            - type: string
        email:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        isGuest:
          type: boolean
        createdAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - isGuest
    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [subscription, lease, tos]
        status:
          type: string
          enum: [pending, analyzing, complete, error]
        overallScore:
          type: ["string", "null"]
          enum: [safe, caution, red_flag, null]
        clauseCount:
          type: integer
        redFlagCount:
          type: integer
        cautionCount:
          type: integer
        safeCount:
          type: integer
        uploadedAt:
          type: string
          format: date-time
        rawText:
          type: string
        errorMessage:
          type: ["string", "null"]
      required:
        - id
        - title
        - category
        - status
        - clauseCount
        - redFlagCount
        - cautionCount
        - safeCount
        - uploadedAt
        - rawText
    DocumentInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        category:
          type: string
          enum: [subscription, lease, tos]
        rawText:
          type: string
          minLength: 1
      required:
        - title
        - category
        - rawText
    DocumentWithClauses:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          enum: [subscription, lease, tos]
        status:
          type: string
          enum: [pending, analyzing, complete, error]
        overallScore:
          type: ["string", "null"]
          enum: [safe, caution, red_flag, null]
        clauseCount:
          type: integer
        redFlagCount:
          type: integer
        cautionCount:
          type: integer
        safeCount:
          type: integer
        uploadedAt:
          type: string
          format: date-time
        rawText:
          type: string
        errorMessage:
          type: ["string", "null"]
        clauses:
          type: array
          items:
            $ref: "#/components/schemas/Clause"
      required:
        - id
        - title
        - category
        - status
        - clauseCount
        - redFlagCount
        - cautionCount
        - safeCount
        - uploadedAt
        - rawText
        - clauses
    Clause:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        text:
          type: string
        categoryTag:
          type: string
        riskLevel:
          type: string
          enum: [safe, caution, red_flag]
        explanation:
          type: string
        negotiationPrompt:
          type: ["string", "null"]
        positionStart:
          type: integer
        positionEnd:
          type: integer
      required:
        - id
        - documentId
        - text
        - categoryTag
        - riskLevel
        - explanation
        - positionStart
        - positionEnd
    ScanStatus:
      type: object
      properties:
        documentId:
          type: integer
        status:
          type: string
          enum: [pending, analyzing, complete, error]
        progress:
          type: integer
        message:
          type: string
      required:
        - documentId
        - status
        - progress
        - message
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error