# lib/api-zod/src/generated/types/clause.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Fine-Print Guardian API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClauseRiskLevel } from './clauseRiskLevel';

export interface Clause {
  id: number;
  documentId: number;
  text: string;
  categoryTag: string;
  riskLevel: ClauseRiskLevel;
  explanation: string;
  /** @nullable */
  negotiationPrompt?: string | null;
  positionStart: number;
  positionEnd: number;
}