# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Fine-Print Guardian API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentCategory } from './documentCategory';
import type { DocumentOverallScore } from './documentOverallScore';
import type { DocumentStatus } from './documentStatus';

export interface Document {
  id: number;
  title: string;
  category: DocumentCategory;
  status: DocumentStatus;
  /** @nullable */
  overallScore?: DocumentOverallScore;
  clauseCount: number;
  redFlagCount: number;
  cautionCount: number;
  safeCount: number;
  uploadedAt: Date;
  rawText: string;
  /** @nullable */
  errorMessage?: string | null;
}