# lib/api-zod/src/generated/types/documentWithClauses.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Fine-Print Guardian API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Clause } from './clause';
import type { DocumentWithClausesCategory } from './documentWithClausesCategory';
import type { DocumentWithClausesOverallScore } from './documentWithClausesOverallScore';
import type { DocumentWithClausesStatus } from './documentWithClausesStatus';

export interface DocumentWithClauses {
  id: number;
  title: string;
  category: DocumentWithClausesCategory;
  status: DocumentWithClausesStatus;
  /** @nullable */
  overallScore?: DocumentWithClausesOverallScore;
  clauseCount: number;
  redFlagCount: number;
  cautionCount: number;
  safeCount: number;
  uploadedAt: Date;
  rawText: string;
  /** @nullable */
  errorMessage?: string | null;
  clauses: Clause[];
}